+++ Firestore/core/src/util/exception.h
@@ -37,6 +37,7 @@
 
 #include "Firestore/core/src/util/string_format.h"
 #include "absl/base/attributes.h"
+#include "absl/base/config.h"
 
 namespace firebase {
 namespace firestore {